requirements.txt
streamlit
tensorflow==2.19.0
opencv-python-headless==4.11.0.86
Pillow
matplotlib
numpy
